Use a package extension for Enzyme
