+++ Project.toml
@@ -7,18 +7,25 @@ ADTypes = "47edcb42-4c32-4615-8424-f2b9edc5f35b"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 NLPModels = "a4795742-8479-5a88-8948-cc11e1c8c1a6"
-Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 ReverseDiff = "37e2e3b7-166d-5795-8a7a-e32c996b4267"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 SparseConnectivityTracer = "9f842d2f-2579-4b1d-911e-f412cf18a3f5"
 SparseMatrixColorings = "0a514795-09f3-496d-8182-132a7b665d35"
 
+[weakdeps]
+Enzyme = "7da242da-08ed-463a-9acd-ee780be4f1d9"
+
+[extensions]
+ADNLPModelsEnzymeExt = "Enzyme"
+
 [compat]
 ADTypes = "1.2.1"
-ForwardDiff = "0.9, 0.10, 1"
+ForwardDiff = "1"
 NLPModels = "0.21.5"
-Requires = "1"
 ReverseDiff = "1"
-SparseConnectivityTracer = "1.0"
+SparseConnectivityTracer = "1"
 SparseMatrixColorings = "0.4.21"
+Enzyme = "0.13.129"
+LinearAlgebra = "1.10"
+SparseArrays = "1.10"
 julia = "1.10"
