+++ docs/Project.toml
@@ -12,12 +12,11 @@ Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 SolverBenchmark = "581a75fa-a23a-52d0-a590-d6201de2218a"
 SparseConnectivityTracer = "9f842d2f-2579-4b1d-911e-f412cf18a3f5"
 SparseMatrixColorings = "0a514795-09f3-496d-8182-132a7b665d35"
-Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [compat]
 DataFrames = "1"
 Documenter = "1.0"
-ManualNLPModels = "0.1"
+ManualNLPModels = "0.2"
 NLPModels = "0.21.5"
 NLPModelsJuMP = "0.13"
 OptimizationProblems = "0.8"
@@ -26,4 +25,3 @@ Plots = "1"
 SolverBenchmark = "0.6"
 SparseConnectivityTracer = "1.0"
 SparseMatrixColorings = "0.4.21"
-Zygote = "0.6.62"
