+++ src/nls.jl
@@ -71,7 +71,6 @@ One can specify a new backend with the keyword arguments `backend::ADNLPModels.A
 There are three pre-coded backends:
 - the default `ForwardDiffAD`.
 - `ReverseDiffAD`.
-- `ZygoteDiffAD` accessible after loading `Zygote.jl` in your environment.
 For an advanced usage, one can define its own backend and redefine the API as done in [ADNLPModels.jl/src/forward.jl](https://github.com/JuliaSmoothOptimizers/ADNLPModels.jl/blob/main/src/forward.jl).
 
 # Examples
@@ -83,9 +82,6 @@ x0 = ones(3)
 nvar = 3
 ADNLSModel(F, x0, nequ) # uses the default ForwardDiffAD backend.
 ADNLSModel(F, x0, nequ; backend = ADNLPModels.ReverseDiffAD) # uses ReverseDiffAD backend.
-
-using Zygote
-ADNLSModel(F, x0, nequ; backend = ADNLPModels.ZygoteAD)
 ```
 
 ```julia
@@ -97,9 +93,6 @@ c(x) = [1x[1] + x[2]; x[2]]
 nvar, ncon = 3, 2
 ADNLSModel(F, x0, nequ, c, zeros(ncon), zeros(ncon)) # uses the default ForwardDiffAD backend.
 ADNLSModel(F, x0, nequ, c, zeros(ncon), zeros(ncon); backend = ADNLPModels.ReverseDiffAD) # uses ReverseDiffAD backend.
-
-using Zygote
-ADNLSModel(F, x0, nequ, c, zeros(ncon), zeros(ncon); backend = ADNLPModels.ZygoteAD)
 ```
 
 For in-place constraints and residual function, use one of the following constructors:
