+++ test/Project.toml
@@ -11,10 +11,10 @@ SparseMatrixColorings = "0a514795-09f3-496d-8182-132a7b665d35"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [compat]
-ForwardDiff = "0.10"
-ManualNLPModels = "0.1"
-NLPModels = "0.21"
+ForwardDiff = "1"
+ManualNLPModels = "0.2"
+NLPModels = "0.21.5"
 NLPModelsModifiers = "0.7"
 NLPModelsTest = "0.10"
 ReverseDiff = "1"
-SparseMatrixColorings = "0.4.0"
+SparseMatrixColorings = "0.4.21"
